Logging out action again, explicitly cast I2C analog value (scaled to 0-100) to int

+++ analog-to-I2C/analog-to-I2C.ino
@@ -26,7 +26,7 @@ int mean = 0;
 // calling functions from here does not seem to work, so using only global variables
 void answer_request()
 {  
-  TinyWireS.send(mean*100/255);
+  TinyWireS.send((int)(mean*100/255));
 }
 
 void setup() {
